HexTile Initial Commit
Since you're working on the module, I figured I'd add this to the Repo.
The grid isn't implemented yet, but we could always collaborate or I
could share my notes if you run into problems.

The only thing you'll need to add to this code is the sound that will
play when the tile is toggled on or off, but the rest of the code should
work fine, even in Android.